requirements.txt
asgiref==3.4.1
Django==3.2.9
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
psycopg2==2.9.1
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
celery
redis